Improve book read logic to track reads over pick up/put down events.
Previously we considered a new reading event to start every time the book was
picked up. Now if the book is picked up at the last position it was put down
on, and then proceeded in a forwards direction, we consider it a single reading
event that was just interrupted by something.

Several minor refactors were also added to support this, primarily when a new
event is recorded, its position defaults to the position of the prior event.