Allow backwards jumps across files.
This was only ever a sanity check that I'd never hit, now I have log examples
that start at the epoch and this check prevents their processing. Treating them
the same was we treat jumps back to the epoch within a file seems to work out
fine, so I see no need to keep this overly cautious check.